Restructure project folders
In order to handle that the top-level folder might not always
have the same name the library is now in a sub-folder and the
addon path is added to the python import path in register().

This also prevents the library from being imported when the
addon is disabled, which was possible before.